+++ python/tests_old/algorithm/test_explode.py
