+++ python/tests_old/core/test_array.py
