+++ python/tests_old/core/test_geometry_type.py
